refactor: Moved whole login screen UI code to LoginScreen.kt file.
