+++ composeApp/src/wasmJsMain/resources/config.dev.json
@@ -1,6 +1,6 @@
 {
   "clientId": "248107412465-i64fdf66a6f4nrj7232ghdmvbsg91pp3.apps.googleusercontent.com",
-  "redirectUri": "http://localhost:8081/",
+  "redirectUri": "http://localhost:8080/",
   "getNumberOfPagesUrl": "http://localhost:5109/api/cars/getCountPages",
   "getCarsFromPageUrl": "http://localhost:5109/api/cars/getPage",
   "googleAuthUrl": "https://accounts.google.com/o/oauth2/v2/auth",
